Rename : pwToggle to pwTypeToggle
